Change World cell weights to double
- Updated World class to store cell weights as double instead of int
- Changed constants FREE and BLOCK to double (1.0 and -1.0)
- Updated getWeight and setWeight signatures to use double
- Updated tests to check double weights using helper function for precision

+++ src/world.cpp
@@ -3,7 +3,7 @@
 
 /***************** CONSTRUCTOR *****************/
 
-World::World(int w, int h) : width(w), height(h), grid(h, std::vector<int>(w, FREE)) {}
+World::World(int w, int h) : width(w), height(h), grid(h, std::vector<double>(w, FREE)) {}
 
 
 /****************** IS BOUNDS ******************/
@@ -45,7 +45,7 @@ double World::getWeight(const State& s) const
 
 /***************** SET WEIGHT ****************/
 
-bool World::setWeight(const State& s, int weight)
+bool World::setWeight(const State& s, double weight)
 {
     if (!inBounds(s.x, s.y))
     {
